Tell the kernel to boot from mmc1 rather then mmc0
With our old kernel (4.2), mmc device nodes could be different wether a
card was in the slot or not. E.g. our eMMC, which sits on the second
mmc controller, could be mmc0 or mmc1 depending on whether an SD card was
found.

With the newer kernel, this was changed to depend on the availability of
the slot, and not the card. E.g. if there are 2 devices defined in the
devicetree, we now get 2 mmc0 and mmc1, even if mmc0 does not contain a
card. This makes the emmc in our case predictable at mmc1.

Signed-off-by: Olliver Schinagl <o.schinagl@ultimaker.com>